feat: Add hotfix workflow for Workshop
This commit adds a new workflow file, `hotfix.yml`, to handle hotfixes for the Workshop. The workflow is triggered when a pull request is closed on the `main` branch. It checks if the pull request description contains the keyword '`hotfix`' using the `pr-comment-checker` action. If the condition is met, it runs the `AnarkisGaming/workshop` action to upload the changes to the Workshop.